Fixes to Chapter View
Chapter Reader is now an own Activity, fixes a lot of problems
