+++ app/src/main/AndroidManifest.xml
@@ -19,6 +19,11 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".ChapterReader">
+
+        </activity>
+
     </application>
 
 </manifest>
