+++ local.properties
@@ -7,5 +7,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Tue Sep 15 00:48:48 CEST 2015
+#Tue Sep 15 16:56:44 CEST 2015
 sdk.dir=G\:\\Android\\sdk
